fix extraneous bundling of AppRouter
