+++ packages/next/src/client/components/router-reducer/fetch-server-response.ts
@@ -25,7 +25,6 @@ import {
   RSC_CONTENT_TYPE_HEADER,
   NEXT_DID_POSTPONE_HEADER,
 } from '../app-router-headers'
-import { urlToUrlWithoutFlightMarker } from '../app-router'
 import { callServer } from '../../app-call-server'
 import { PrefetchKind } from './router-reducer-types'
 import { hexHash } from '../../../shared/lib/hash'
@@ -36,6 +35,23 @@ export type FetchServerResponseResult = [
   postponed?: boolean,
   intercepted?: boolean,
 ]
+
+function urlToUrlWithoutFlightMarker(url: string): URL {
+  const urlWithoutFlightParameters = new URL(url, location.origin)
+  urlWithoutFlightParameters.searchParams.delete(NEXT_RSC_UNION_QUERY)
+  if (process.env.NODE_ENV === 'production') {
+    if (
+      process.env.__NEXT_CONFIG_OUTPUT === 'export' &&
+      urlWithoutFlightParameters.pathname.endsWith('.txt')
+    ) {
+      const { pathname } = urlWithoutFlightParameters
+      const length = pathname.endsWith('/index.txt') ? 10 : 4
+      // Slice off `/index.txt` or `.txt` from the end of the pathname
+      urlWithoutFlightParameters.pathname = pathname.slice(0, -length)
+    }
+  }
+  return urlWithoutFlightParameters
+}
 
 function doMpaNavigation(url: string): FetchServerResponseResult {
   return [urlToUrlWithoutFlightMarker(url).toString(), undefined, false, false]
